RoutineServiceProxy: creates the class
This commits adds the files routine-service-proxy.hpp and its .cpp
with the declaration and definition of the class RoutineServiceProxy and
its states. Just to be more explicit, RoutineServiceProxy is the old
routineService::Proxy class with some improvements.

Some of news are the adoption of modern iterators (range based "fors")
"for(auto& [key, value] : thingtobeiterated)" and the adoption of
smart pointers.