StaticServiceProxy: creates the class
This commits adds the files static-service-proxy.hpp and its .cpp
with the declaration and definition of the class StaticServiceProxy and
its states. Just to be more explicit, StaticServiceProxy is the old
staticService::Proxy class with some improvements.

Some of news are the adoption of modern iterators (range based "fors")
"for(auto& [key, value] : thingToBeIterated)" and the adoption
of smart pointers.